Add new shopping card for Redux and refractor code into new folder
