+++ src/pages/ShoppingListsMain/index.tsx
@@ -13,6 +13,7 @@ const ShoppingListMain = () => {
       <GroupCardsContainer>
         <ShoppingCard names="name1, name2, name4" onClick={() => history.push(PATHS.EXAMPLE_SHOPPING_PAGE)} />
         <ShoppingCard names="zhikai, shaokiat" onClick={() => history.push(PATHS.SHOPPING_PAGE)} />
+        <ShoppingCard names="zhikai, shaokiat (Redux)" onClick={() => history.push(PATHS.PAIR6_SHOPPING_PAGE)} />
         {/* Create group cards here */}
       </GroupCardsContainer>
     </>
