+++ src/pages/SkZkShoppingPageRedux/styles/ShoppingPageRedux.styled.ts
@@ -1,0 +1,36 @@
+import styled from 'styled-components'
+
+export const ShoppingPageDiv = styled.div``
+
+export const ShoppingListHeader = styled.h1`
+  text-align: center;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  font-weight: 600;
+  font-size: 3rem;
+  margin: 20px auto;
+`
+
+export const ItemRowContainer = styled.div`
+  width: 80vw;
+  margin: auto;
+  justify-content: center;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+`
+
+export const AddTextInput = styled.input`
+  border-color: #000000;
+  width: 20vw;
+  margin: 10px;
+  padding: 10px;
+  font-size: 1rem;
+  align-items: center;
+`
+
+export const CartTable = styled.table`
+  margin: auto;
+  width: 50vw;
+`
