+++ src/store/shoppingLists/actions.ts
@@ -1,5 +1,5 @@
 import { Dispatch, GetState } from '../types'
-import { ActionTypes, SHOPPING_LIST_ACTIONS, ExampleType, CartItem } from './types'
+import { ActionTypes, SHOPPING_LIST_ACTIONS, ExampleType } from './types'
 
 //set my 'exampleList' state with my argument 'newList'
 //to call this in your component or page, wrap with dispatch
@@ -13,32 +13,3 @@ export const mockActionSetMyExampleList =
       exampleList: newList ?? exampleList, // if newList is undefined, use exampleList
     })
   }
-
-export const SetShoppingCart = (newCart: CartItem[]) => async (dispatch: Dispatch<ActionTypes>, getState: GetState) => {
-  const { shoppingCart } = getState().shoppingLists
-  dispatch({
-    type: SHOPPING_LIST_ACTIONS.SET_SHOPPING_CART,
-    shoppingCart: newCart ?? shoppingCart,
-  })
-}
-
-export const AddItemQuantity = (item: CartItem) => async (dispatch: Dispatch<ActionTypes>) => {
-  dispatch({
-    type: SHOPPING_LIST_ACTIONS.ADD_ITEM_QUANTITY,
-    update: item,
-  })
-}
-
-export const MinusItemQuantity = (item: CartItem) => async (dispatch: Dispatch<ActionTypes>) => {
-  dispatch({
-    type: SHOPPING_LIST_ACTIONS.MINUS_ITEM_QUANTITY,
-    update: item,
-  })
-}
-
-export const DeleteCartItem = (item: CartItem) => async (dispatch: Dispatch<ActionTypes>) => {
-  dispatch({
-    type: SHOPPING_LIST_ACTIONS.DELETE_CART_ITEM,
-    update: item,
-  })
-}
