+++ src/store/store.ts
@@ -3,9 +3,11 @@ import thunk from 'redux-thunk'
 import { composeWithDevTools } from 'redux-devtools-extension'
 
 import { shoppingLists } from './shoppingLists/reducer'
+import { shoppingListsSKZK } from './pair6/reducer'
 
 export const rootReducer = combineReducers({
   shoppingLists,
+  shoppingListsSKZK,
 })
 const middlewares = [thunk]
 const store = createStore(rootReducer, composeWithDevTools(applyMiddleware(...middlewares)))
